libnm: add NM_OBJECT_CLIENT property
We have nm_object_get_client() property that returns a reference
to the NMClient instance. This is actually useful, because if
the function returns %NULL, it means that the object was removed
from the cache.

On the other hand, the user cannot subscribe to notifications when this
happens. Well, there are otherwise pointless signals like
NM_CLIENT_DEVICE_REMOVED, which we wouldn't need if we had a general
mechanism for NMObject instances.

Add a GObject property "client", which is just that mechanism.

https://gitlab.freedesktop.org/NetworkManager/NetworkManager/-/merge_requests/902