core: fix `nmcli device connect dummy0` to add-and-activate dummy profile
   $ ip link add dd type dummy

   $ nmcli device
   DEVICE             TYPE      STATE                   CONNECTION
   ...
   dd                 dummy     unmanaged               --

   $ nmcli device connect dd
   Error: Failed to add/activate new connection: A 'dummy' setting is required.

There are two problems here. The first is that we don't pass
the interface name to nm_utils_complete_generic(), but dummy
devices require "connection.interface-name" set. As a consequence,
nm_utils_complete_generic() fails to normalize the connection
and there is no [dummy] setting. Which then results in a failure
with complete_connection().

The important part of the fix is to set the interface name. Once
we do that, nm_utils_complete_generic() should be able to add
the [dummy] setting and the second part is not strictly necessary.
Still, the job of complete_connection() is not to verify the
profile but to create it with best effort. If a [dummy] setting
is still missing, we should just add it. The caller will then
again try to normalize/verify the connection, and that might then
fail -- but this time not with the wrong error message about
missing 'dummy' setting.

https://bugzilla.redhat.com/show_bug.cgi?id=1763054